% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dir_suffixes.R
\name{check_dir_suffixes}
\alias{check_dir_suffixes}
\title{Check if all directories end with specified suffixes.}
\usage{
check_dir_suffixes(dirs, suffix)
}
\arguments{
\item{dirs}{A character vector of directory names.}

\item{suffix}{A character vector of suffixes to check.}
}
\value{
Invisibly returns TRUE if all directories match the suffixes, FALSE otherwise.
}
\description{
This is an internal function which checks if all directories in a given list end with the specified suffixes.
}
