% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlineCount.R
\name{longlineCount}
\alias{longlineCount}
\title{Process longline catch data from CSV files in a directory.}
\usage{
longlineCount(inDir)
}
\arguments{
\item{inDir}{The path to the directory containing the CSV files.}
}
\value{
A data frame containing the combined and processed data from all CSV files.
}
\description{
This function reads CSV files from a specified directory, standardizes the data,
extracts information from the file names, fixes the dataframe structure,
and combines the data into a single data frame.
}
\examples{
\dontrun{
# Assuming you have CSV files in a directory named "data"
data <- longlineCount("C:/Users/BirdLife/Documents/Data")
View(data)
}

}
