% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_values_exclude_first.R
\name{standardize_values_exclude_first}
\alias{standardize_values_exclude_first}
\title{Standardize values in a data frame, excluding the first column.}
\usage{
standardize_values_exclude_first(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with standardized values.
}
\description{
TThis is an internal function which standardizes the values in a data frame, excluding the first column.
It handles character and numeric values, replacing missing values and standardizing text.
}
